gemini_added

+++ routers/todo.py
@@ -1,18 +1,28 @@
 from starlette import status
 from pydantic import BaseModel, Field
-from fastapi import FastAPI, Depends, Path, Query, HTTPException, APIRouter
+from fastapi import FastAPI, Depends, Path, Query, HTTPException, APIRouter , Request
 from sqlalchemy.orm import Session
 from models import Base, Todo
 from database import engine, SessionLocal
 from typing import Annotated
 from routers.auth import get_current_user
+from starlette.responses import RedirectResponse
+from fastapi.templating import Jinja2Templates
+from dotenv import load_dotenv
+import google.generativeai as genai
+import os
+from langchain_google_genai import ChatGoogleGenerativeAI
+from langchain_core.messages import HumanMessage, AIMessage
+import markdown
+from bs4 import BeautifulSoup
+
 
 router = APIRouter(
     prefix="/todo",
     tags = ["Todo"]
 )
 
-
+templates_  = Jinja2Templates(directory="templates")
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
     description: str = Field(min_length=3 , max_length= 1000)
@@ -30,36 +40,80 @@ def get_db():
 db_dependency = Annotated[Session, Depends(get_db)] #dependency = bağımlılık , annotated = dipnot
 user_dependency = Annotated[dict, Depends(get_current_user)]
 
+def redirect_to_login():
+    redirect_response = RedirectResponse(url="/auth/login-page", status_code=status.HTTP_302_FOUND)
+    redirect_response.delete_cookie("acces_token")#kullanıcının coociesi geçersiz ise siliniyor.
+    return redirect_response
+
+
+@router.get("/todo-page")
+async def render_todo_page(request: Request, db: db_dependency):
+    try:
+        user = await get_current_user(request.cookies.get('access_token'))
+        if user is None:
+            return redirect_to_login()
+        todos = db.query(Todo).filter(Todo.owner_id == user.get("id")).all()
+        return templates_.TemplateResponse("todo.html",{"request":request,"todos":todos, "user": user})
+    except:
+        return redirect_to_login()
+
+@router.get("/add-todo-page")
+async def render_add_todo_page(request: Request):
+    try:
+        user = await get_current_user(request.cookies.get('access_token'))
+        if user is None:
+            return redirect_to_login()
+        return templates_.TemplateResponse("add-todo.html",{"request":request, "user": user})
+    except:
+        return redirect_to_login()
+
+
+@router.get("/edit-todo-page/{todo_id}")
+async def render_edit_todo_page(request: Request,todo_id:int, db: db_dependency):
+    try:
+        user = await get_current_user(request.cookies.get('access_token'))
+        if user is None:
+            return redirect_to_login()
+        todo = db.query(Todo).filter(Todo.id == todo_id).first()
+        return templates_.TemplateResponse("edit-todo.html",{"request":request,"todo":todo, "user": user})
+    except:
+        return redirect_to_login()
+
+
 @router.get("/read_all")
 async def read_all(user: user_dependency,db: db_dependency):
     if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
-    return db.query(Todo).all()
+    return db.query(Todo).filter(Todo.owner_id == user.get('id')).all()
 
 
-@router.get("/read_one/{todo_id}", status_code =status.HTTP_200_OK)
-async def read_by_id(db: db_dependency, todo_id: int = Path(gt = 0)):
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()  #first ilk gelen elemanı döndürecek
+@router.get("/todo/{todo_id}", status_code =status.HTTP_200_OK)
+async def read_by_id(user: user_dependency ,db: db_dependency, todo_id: int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()#first ilk gelen elemanı döndürecek
     if todo is not None:
         return todo
     raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, detail = "todo not found" )
 
 
-@router.post("/create", status_code=status.HTTP_201_CREATED)
-async def create(user: user_dependency,db:db_dependency, todo_request: TodoRequest):
-    if user in None:
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create(user: user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
 
-    todo = Todo(**todo_request.model_dump(), owner_id =user.get("id"))
+    todo = Todo(**todo_request.model_dump(), owner_id=user.get("id"))
+    todo.description = create_todo_with_gemini(todo.description)
     db.add(todo)
     db.commit()#add dedikten sonra commit demez isek çalışmaz. işlemin yapılacağı anlamına gelir
 
-@router.put("/update/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update(db:db_dependency,
+@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update(db:db_dependency ,user: user_dependency,
                    todo_request: TodoRequest,
                    todo_id: int = Path(gt=0)):
-
-    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(Todo.id == todo_id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
     todo.title = todo_request.title
@@ -68,11 +122,34 @@ async def update(db:db_dependency,
     todo.complete = todo_request.complete
     db.add(todo)
     db.commit()
-@router.delete("/delete/{todo_id}", status_code=status.HTTP_200_OK)
-async def delete(db:db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todo).filter(todo_id == Todo.id).first()
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_200_OK)
+async def delete(user: user_dependency,db:db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    todo = db.query(Todo).filter(todo_id == Todo.id).filter(Todo.owner_id == user.get('id')).first()
     if todo is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
     '''db.query(Todo).filter(Todo.id == todo_id).delete() =  silme yöntemlerinden birisi. garanticiler bunu kullanabiliyor :)'''
     db.delete(todo)
     db.commit()
+
+def markdown_to_text(markdown_string):
+    html = markdown.markdown(markdown_string)
+    soup = BeautifulSoup(html,"html.parser")
+    text = soup.getText()
+    return text
+
+def create_todo_with_gemini(todo_string: str):
+    load_dotenv()
+    genai.configure(api_key=os.environ.get("GOOGLE_API_KEY"))
+    llm = ChatGoogleGenerativeAI(model="gemini-pro")
+    response = llm.invoke(
+    [
+            HumanMessage(
+                content="I will provide you a todo item to add my to do list. What i want you to do is to create a longer and more comprehensive description of that  todo item, my next message will be my todo. Ve bütün bunları türkçe olarak cevapla."),
+            HumanMessage(content=todo_string)
+    ]
+    )
+    return response.content
+
+
